+++ packages/cli/src/util/createServer.ts
@@ -11,6 +11,7 @@ import * as split from 'split2';
 import { PassThrough, Readable } from 'stream';
 import { LOG_COLOR_MAP } from '../const/options';
 import { createExamplePath } from './paths';
+import { attachTagsToParamsValues, transformPathParamsValues } from './colorizer';
 
 async function createMultiProcessPrism(options: CreateBaseServerOptions) {
   if (cluster.isMaster) {
@@ -78,13 +79,14 @@ async function createPrismServerWithLogger(options: CreateBaseServerOptions, log
 
   operations.forEach(resource => {
     const path = pipe(
-      createExamplePath(resource),
+      createExamplePath(resource, attachTagsToParamsValues),
       Either.getOrElse(() => resource.path)
     );
 
-    logInstance.note(`${resource.method.toUpperCase().padEnd(10)} ${address}${path}`);
+    logInstance.note(
+      `${resource.method.toUpperCase().padEnd(10)} ${address}${transformPathParamsValues(path, chalk.bold.cyan)}`
+    );
   });
-
   logInstance.start(`Prism is listening on ${address}`);
 
   return server;